FIX canvas start timeout error, EDIT readme
Added an environment variable in docker-compose.yml to increase the
Passenger service timeout to 5 minutes. When I tried starting Canvas for
the first time, it always takes longer than the default timeout, so I
keep seeing the Passenger error page. Increasing the timeout finally
allowed me to see the Canvas login page.

Added a section in the readme on how to build an updated Canvas image on
your own since the official docker image hasn't been updated for some
time.